Add code review changes: download apictl from github use the direct project property to refer target dir in pom remove new URLs file and move content to utils get apictl version from pom using system properties Remove souts Move cors_apiDeploy to before tests
