+++ integration/mock-backend-server/pom.xml
@@ -19,7 +19,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <parent>
-        <artifactId>micro-gateway-tests</artifactId>
+        <artifactId>integration</artifactId>
         <groupId>org.wso2.am.microgw</groupId>
         <version>4.0.0-m10-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
