+++ integration/test-integration/src/test/java/org/wso2am/micro/gw/tests/testCases/jwtGenerator/JwtGeneratorTestCase.java
@@ -28,7 +28,7 @@ import org.wso2am.micro.gw.tests.util.HttpsClientRequest;
 import org.wso2am.micro.gw.tests.util.HttpResponse;
 import org.wso2am.micro.gw.tests.util.TestConstant;
 import org.wso2am.micro.gw.tests.util.TokenUtil;
-import org.wso2am.micro.gw.tests.util.URLs;
+import org.wso2am.micro.gw.tests.util.Utils;
 
 import java.util.Base64;
 import java.util.HashMap;
@@ -53,7 +53,7 @@ public class JwtGeneratorTestCase {
         //test endpoint with token
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtTokenProd);
         HttpResponse response = HttpsClientRequest
-                .doGet(URLs.getServiceURLHttps("v2/jwtheader"), headers);
+                .doGet(Utils.getServiceURLHttps("v2/jwtheader"), headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getData(), ResponseConstants.VALID_JWT_RESPONSE);
         Assert.assertEquals(response.getResponseCode(), 200, "Response code mismatched");
@@ -64,7 +64,7 @@ public class JwtGeneratorTestCase {
         Map<String, String> headers = new HashMap<>();
         //test endpoint with token
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtTokenProd);
-        HttpResponse response = HttpsClientRequest.doGet(URLs.getServiceURLHttps(
+        HttpResponse response = HttpsClientRequest.doGet(Utils.getServiceURLHttps(
                 "/v2/jwttoken") , headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getResponseCode(), 200, "Response code mismatched");
