+++ integration/test-integration/src/test/java/org/wso2am/micro/gw/tests/testCases/jwtValidator/JwtTestCase.java
@@ -27,7 +27,7 @@ import org.wso2am.micro.gw.tests.util.HttpResponse;
 import org.wso2am.micro.gw.tests.util.HttpsClientRequest;
 import org.wso2am.micro.gw.tests.util.TestConstant;
 import org.wso2am.micro.gw.tests.util.TokenUtil;
-import org.wso2am.micro.gw.tests.util.URLs;
+import org.wso2am.micro.gw.tests.util.Utils;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -56,7 +56,7 @@ public class JwtTestCase {
         // Set header
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtWithoutScope);
-        HttpResponse response = HttpsClientRequest.doGet(URLs.getServiceURLHttps(
+        HttpResponse response = HttpsClientRequest.doGet(Utils.getServiceURLHttps(
                 "/v2/pet/2") , headers);
 
         Assert.assertNotNull(response);
@@ -69,7 +69,7 @@ public class JwtTestCase {
         // Set header
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + TestConstant.INVALID_JWT_TOKEN);
-        HttpResponse response = HttpsClientRequest.doGet(URLs.getServiceURLHttps(
+        HttpResponse response = HttpsClientRequest.doGet(Utils.getServiceURLHttps(
                 "/v2/pet/2") , headers);
 
         Assert.assertNotNull(response);
@@ -83,7 +83,7 @@ public class JwtTestCase {
         // Set header
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + TestConstant.EXPIRED_JWT_TOKEN);
-        HttpResponse response = HttpsClientRequest.doGet(URLs.getServiceURLHttps(
+        HttpResponse response = HttpsClientRequest.doGet(Utils.getServiceURLHttps(
                 "/v2/pet/2") , headers);
 
         Assert.assertNotNull(response);
