+++ integration/test-integration/src/test/java/org/wso2am/micro/gw/tests/testCases/jwtValidator/ScopeTest.java
@@ -28,7 +28,7 @@ import org.wso2am.micro.gw.tests.util.HttpClientRequest;
 import org.wso2am.micro.gw.tests.util.HttpResponse;
 import org.wso2am.micro.gw.tests.util.TestConstant;
 import org.wso2am.micro.gw.tests.util.TokenUtil;
-import org.wso2am.micro.gw.tests.util.URLs;
+import org.wso2am.micro.gw.tests.util.Utils;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -54,7 +54,7 @@ public class ScopeTest {
     public void testScopeProtectedResourceInvalidJWT() throws Exception {
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtWithoutScope);
-        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(URLs.getServiceURLHttps(
+        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(Utils.getServiceURLHttps(
                 "/v2/pets/findByTags"), headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getResponseCode(), HttpStatus.SC_FORBIDDEN, "Response code mismatched");
@@ -67,7 +67,7 @@ public class ScopeTest {
     public void testScopeProtectedResourceValidJWT() throws Exception {
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtWithScope);
-        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(URLs.getServiceURLHttps("/v2/pet/findByStatus"), headers);
+        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(Utils.getServiceURLHttps("/v2/pet/findByStatus"), headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getResponseCode(), HttpStatus.SC_OK, "Response code mismatched");
         Assert.assertEquals(response.getData(), ResponseConstants.RESPONSE_BODY,
@@ -78,7 +78,7 @@ public class ScopeTest {
     public void testMultipleScopeProtectedResourceValidJWT() throws Exception {
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtWithScope);
-        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(URLs.getServiceURLHttps("/v2/pets/findByTags"), headers);
+        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(Utils.getServiceURLHttps("/v2/pets/findByTags"), headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getResponseCode(), HttpStatus.SC_OK, "Response code mismatched");
         Assert.assertEquals(response.getData(), ResponseConstants.PET_BY_ID_RESPONSE,
@@ -89,7 +89,7 @@ public class ScopeTest {
     public void testMultipleScopeProtectedResourceValidMultiScopeJWT() throws Exception {
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtWithMultipleScopes);
-        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(URLs.getServiceURLHttps("/v2/pets/findByTags"), headers);
+        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(Utils.getServiceURLHttps("/v2/pets/findByTags"), headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getResponseCode(), HttpStatus.SC_OK, "Response code mismatched");
         Assert.assertEquals(response.getData(), ResponseConstants.PET_BY_ID_RESPONSE,
@@ -100,7 +100,7 @@ public class ScopeTest {
     public void testMultipleScopeProtectedResourceInvalidMultiScopeJWT() throws Exception {
         Map<String, String> headers = new HashMap<String, String>();
         headers.put(HttpHeaderNames.AUTHORIZATION.toString(), "Bearer " + jwtWithMultipleInvalidScopes);
-        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(URLs.getServiceURLHttps("/v2/pets/findByTags"), headers);
+        HttpResponse response = HttpClientRequest.retryGetRequestUntilDeployed(Utils.getServiceURLHttps("/v2/pets/findByTags"), headers);
         Assert.assertNotNull(response);
         Assert.assertEquals(response.getResponseCode(), HttpStatus.SC_FORBIDDEN,"Response code mismatched");
         Assert.assertTrue(
