--- integration/test-integration/src/test/java/org/wso2am/micro/gw/tests/util/URLs.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.wso2am.micro.gw.tests.util;
-
-import java.net.MalformedURLException;
-import java.net.URL;
-
-public class URLs {
-    public static String getAdapterServiceURLHttps(String servicePath) throws MalformedURLException {
-        return new URL(new URL("https://localhost:" + TestConstant.ADAPTER_PORT), servicePath)
-                .toString();
-    }
-
-    public static String getServiceURLHttps(String servicePath) throws MalformedURLException {
-        return new URL(new URL("https://localhost:" + TestConstant.GATEWAY_LISTENER_HTTPS_PORT), servicePath)
-                .toString();
-    }
-
-    public static String getMockServiceURLHttp(String servicePath) throws MalformedURLException {
-        return new URL(new URL("http://localhost:" + TestConstant.MOCK_SERVER_PORT), servicePath).toString();
-    }
-}
